requirements.txt
pandas
numpy
ta
python-dotenv
requests 